requirements.txt
alabaster==1.0.0
anyio==4.12.1
apeye==1.4.1
apeye-core==1.1.5
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.11.1
asttokens==3.0.1
async-lru==2.1.0
attrs==25.4.0
autodocsumm==0.2.14
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.3.0
breathe==4.36.0
CacheControl==0.14.4
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cssselect2==0.9.0
cssutils==2.11.1
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
dict2css==0.3.0.post1
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docutils==0.21.2
domdf_python_tools==3.10.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.25.0
fonttools==4.61.1
fqdn==1.5.1
fsspec==2026.2.0
greenlet==3.3.2
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imagesize==2.0.0
importlib_metadata==8.7.1
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-cache==1.0.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lark==1.3.1
latexcodec==3.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdit-py-plugins==0.5.0
mdurl==0.1.2
mistune==3.2.0
more-itertools==10.8.0
msgpack==1.1.2
myst-nb==1.4.0
myst-parser==5.0.0
natsort==8.4.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.5.3
notebook_shim==0.2.4
numpy==2.4.2
overrides==7.7.0
packaging==26.0
pandas==3.0.0
pandocfilters==1.5.1
parso==0.8.5
pillow==12.1.0
platformdirs==4.5.1
prometheus_client==0.24.1
prompt_toolkit==3.0.52
psutil==7.2.2
psycopg2-binary==2.9.11
pure_eval==0.2.3
pybtex==0.25.1
pybtex-docutils==1.0.3
pycparser==3.0
pyenchant==3.3.0
pygame==2.6.1
Pygments==2.19.2
PyJWT==2.12.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pywinpty==3.0.3
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
roman==5.2
roman-numerals==4.1.0
roman-numerals-py==4.1.0
rpds-py==0.30.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
seaborn==0.13.2
Send2Trash==2.1.0
six==1.17.0
snowballstemmer==3.0.1
soupsieve==2.8.3
Sphinx==8.2.3
sphinx-autodoc-typehints==3.5.2
sphinx-jinja2-compat==0.4.1
sphinx-prompt==1.10.2
sphinx-rtd-theme==3.0.2
sphinx-tabs==3.4.5
sphinx-toolbox==4.1.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-bibtex==2.6.5
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sphinxcontrib-spelling==8.0.2
sphinxcontrib-svg2pdfconverter==2.1.0
SQLAlchemy==2.0.48
sqlparse==0.5.5
stack-data==0.6.3
tabulate==0.10.0
terminado==0.18.1
tinycss2==1.4.0
tornado==6.5.4
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
widgetsnbextension==4.0.15
yt-dlp==2026.3.17
zipp==3.23.0
